Move gem name loader to proper namespaced location.
See: https://guides.rubygems.org/name-your-gem/
