+++ lib/manageiq/appliance_console.rb
